feat(daytona): merge hybrid PTY reconnection into baseline process manager
Consolidate the hybrid process manager's PTY reconnection capability into
the baseline DaytonaProcessManager, eliminating code duplication. The
baseline now discovers externally-spawned PTY sessions in list() and
reconnects to them via get() fallback.
